+++ src/prepare-tx.js
@@ -1,13 +1,14 @@
 import {Buffer} from 'safe-buffer';
-import MinterTx, {MinterTxSignature, formatCoin} from 'minterjs-tx';
+import {MinterTx, MinterTxSignature, formatCoin} from 'minterjs-tx';
 import {toHexString} from './utils';
 
 /**
  * @typedef {Object} TxParams
  * @property {string|Buffer} privateKey
  * @property {number} [nonce]
- * @property {number} [gasPrice]
- * @property {string} [gasCoin]
+ * @property {Number} [chainID=1]
+ * @property {number} [gasPrice=1]
+ * @property {string} [gasCoin='BIP']
  * @property {string|Buffer} txType
  * @property {Buffer} txData
  * @property {string} [message]
@@ -18,13 +19,18 @@ import {toHexString} from './utils';
  * @param {TxParams} txParams
  * @return {MinterTx}
  */
-export default function prepareSignedTx(txParams) {
-    const {privateKey, nonce = 1, gasPrice = 1, gasCoin = 'BIP', txType, txData, message} = txParams;
+export default function prepareSignedTx(txParams = {}) {
+    const {privateKey, nonce, chainID = 1, gasPrice = 1, gasCoin = 'BIP', txType, txData, message} = txParams;
+    // throw on falsy nonce except 0
+    if (!nonce && typeof nonce !== 'number') {
+        throw new Error('Invalid nonce specified, tx can\'t be prepared');
+    }
     // @TODO asserts
     const privateKeyBuffer = typeof privateKey === 'string' ? Buffer.from(privateKey, 'hex') : privateKey;
 
     const txProps = {
         nonce: `0x${toHexString(nonce)}`,
+        chainID: `0x${toHexString(chainID)}`,
         gasPrice: `0x${toHexString(gasPrice)}`,
         gasCoin: formatCoin(gasCoin),
         type: txType,