+++ test/issue-check.test.js
@@ -7,13 +7,14 @@ describe('issueCheck()', () => {
         privateKey: '2919c43d5c712cae66f869a524d9523999998d51157dc40ac4d8d80a7602ce02',
         passPhrase: 'pass',
         nonce: 1,
+        chainID: 1,
         coinSymbol: 'MNT',
         value: 10,
         dueBlock: 999999,
     };
     test('should work', () => {
         const check = issueCheck(checkParams);
-        expect(check).toEqual('Mcf89f01830f423f8a4d4e5400000000000000888ac7230489e80000b841ada7ad273bef8a1d22f3e314fdfad1e19b90b1fe8dc7eeb30bd1d391e89af8642af029c138c2e379b95d6bc71b26c531ea155d9435e156a3d113a14c912dfebf001ca0781a7b7d781634bcf632579b99d583887ab093dfbd50b65de5c0e5813028a277a071272d8e1be721f5307f40f87daa4ab632781640f18fd424839396442cc7ff17');
+        expect(check).toEqual('Mcf8a00101830f423f8a4d4e5400000000000000888ac7230489e80000b841e383eff7dc2b162fe6b3af68d1b4a877c629693275e58010cd87f419b24c34bd31e66d8bb36fc0005be3d7f4a446f8bb56c32c738a241c1cb02337d86398b5e7001ca0b053a8ee3b9e32545aa13c77173abc91f5edc07a60695f85b00d94003e27459ea06835ad40bd0cbdc2ae7568eef4f5cbc513ae11d13c465f7ead85e1d2365007bd');
     });
 
     test('should accept buffer private key', () => {
@@ -21,7 +22,7 @@ describe('issueCheck()', () => {
             ...checkParams,
             privateKey: Buffer.from(checkParams.privateKey, 'hex'),
         });
-        expect(check).toEqual('Mcf89f01830f423f8a4d4e5400000000000000888ac7230489e80000b841ada7ad273bef8a1d22f3e314fdfad1e19b90b1fe8dc7eeb30bd1d391e89af8642af029c138c2e379b95d6bc71b26c531ea155d9435e156a3d113a14c912dfebf001ca0781a7b7d781634bcf632579b99d583887ab093dfbd50b65de5c0e5813028a277a071272d8e1be721f5307f40f87daa4ab632781640f18fd424839396442cc7ff17');
+        expect(check).toEqual('Mcf8a00101830f423f8a4d4e5400000000000000888ac7230489e80000b841e383eff7dc2b162fe6b3af68d1b4a877c629693275e58010cd87f419b24c34bd31e66d8bb36fc0005be3d7f4a446f8bb56c32c738a241c1cb02337d86398b5e7001ca0b053a8ee3b9e32545aa13c77173abc91f5edc07a60695f85b00d94003e27459ea06835ad40bd0cbdc2ae7568eef4f5cbc513ae11d13c465f7ead85e1d2365007bd');
     });
 
     test('default dueBlock: 999999', () => {
@@ -31,6 +32,16 @@ describe('issueCheck()', () => {
         })).toEqual(issueCheck({
             ...checkParams,
             dueBlock: 999999999,
+        }));
+    });
+
+    test('default chainID: 1', () => {
+        expect(issueCheck({
+            ...checkParams,
+            chainID: undefined,
+        })).toEqual(issueCheck({
+            ...checkParams,
+            chainID: 1,
         }));
     });